feat(ui): TE-1449 add json and advanced editor to the welcome create alert flow
