+++ thirdeye-ui/src/app/pages/alerts-create-page/alerts-create-simple-page/alerts-create-simple-page.component.tsx
@@ -13,15 +13,23 @@
  * the License.
  */
 import { Grid } from "@material-ui/core";
-import React, { FunctionComponent, useEffect } from "react";
+import React, { FunctionComponent } from "react";
+import { useTranslation } from "react-i18next";
 import { useOutletContext } from "react-router-dom";
 import { AlertDetails } from "../../../components/alert-wizard-v2/alert-details/alert-details.component";
 import { AlertNotifications } from "../../../components/alert-wizard-v2/alert-notifications/alert-notifications.component";
 import { AlertTemplate } from "../../../components/alert-wizard-v2/alert-template/alert-template.component";
+import { Portal } from "../../../components/portal/portal.component";
+import { WizardBottomBar } from "../../../components/welcome-onboard-datasource/wizard-bottom-bar/wizard-bottom-bar.component";
 import { PageContentsGridV1 } from "../../../platform/components";
-import { AlertEditPageOutletContextProps } from "../../alerts-update-page/alerts-update-page.interfaces";
+import {
+    AlertsSimpleAdvancedJsonContainerPageOutletContextProps,
+    BOTTOM_BAR_ELEMENT_ID,
+} from "../../alerts-edit-create-common/alerts-edit-create-common-page.interfaces";
 
 export const AlertsCreateSimplePage: FunctionComponent = () => {
+    const { t } = useTranslation();
+
     const {
         alert,
         handleAlertPropertyChange: onAlertPropertyChange,
@@ -29,37 +37,50 @@ export const AlertsCreateSimplePage: FunctionComponent = () => {
         selectedAlertTemplate,
         setSelectedAlertTemplate,
         alertTemplateOptions,
-        setShowBottomBar,
-    } = useOutletContext<AlertEditPageOutletContextProps>();
-
-    useEffect(() => {
-        setShowBottomBar(true);
-    }, []);
+        isEditRequestInFlight,
+        handleSubmitAlertClick,
+        onPageExit,
+    } = useOutletContext<AlertsSimpleAdvancedJsonContainerPageOutletContextProps>();
 
     return (
-        <PageContentsGridV1>
-            <Grid item xs={12}>
-                <AlertDetails
-                    alert={alert}
-                    onAlertPropertyChange={onAlertPropertyChange}
+        <>
+            <PageContentsGridV1>
+                <Grid item xs={12}>
+                    <AlertDetails
+                        alert={alert}
+                        onAlertPropertyChange={onAlertPropertyChange}
+                    />
+                </Grid>
+                <Grid item xs={12}>
+                    <AlertTemplate
+                        alert={alert}
+                        alertTemplateOptions={alertTemplateOptions}
+                        selectedAlertTemplate={selectedAlertTemplate}
+                        setSelectedAlertTemplate={setSelectedAlertTemplate}
+                        onAlertPropertyChange={onAlertPropertyChange}
+                    />
+                </Grid>
+                <Grid item xs={12}>
+                    <AlertNotifications
+                        alert={alert}
+                        initiallySelectedSubscriptionGroups={[]}
+                        onSubscriptionGroupsChange={onSubscriptionGroupsChange}
+                    />
+                </Grid>
+            </PageContentsGridV1>
+
+            <Portal containerId={BOTTOM_BAR_ELEMENT_ID}>
+                <WizardBottomBar
+                    doNotWrapInContainer
+                    backButtonLabel={t("label.cancel")}
+                    handleBackClick={onPageExit}
+                    handleNextClick={() => handleSubmitAlertClick(alert)}
+                    nextButtonIsDisabled={isEditRequestInFlight}
+                    nextButtonLabel={t("label.create-entity", {
+                        entity: t("label.alert"),
+                    })}
                 />
-            </Grid>
-            <Grid item xs={12}>
-                <AlertTemplate
-                    alert={alert}
-                    alertTemplateOptions={alertTemplateOptions}
-                    selectedAlertTemplate={selectedAlertTemplate}
-                    setSelectedAlertTemplate={setSelectedAlertTemplate}
-                    onAlertPropertyChange={onAlertPropertyChange}
-                />
-            </Grid>
-            <Grid item xs={12}>
-                <AlertNotifications
-                    alert={alert}
-                    initiallySelectedSubscriptionGroups={[]}
-                    onSubscriptionGroupsChange={onSubscriptionGroupsChange}
-                />
-            </Grid>
-        </PageContentsGridV1>
+            </Portal>
+        </>
     );
 };
