+++ thirdeye-ui/src/app/pages/alerts-update-page/alerts-update-simple-page.component.tsx
@@ -13,17 +13,28 @@
  * the License.
  */
 import { Grid } from "@material-ui/core";
-import React, { FunctionComponent, useEffect } from "react";
+import React, { FunctionComponent, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useOutletContext } from "react-router-dom";
 import { AlertDetails } from "../../components/alert-wizard-v2/alert-details/alert-details.component";
 import { AlertNotifications } from "../../components/alert-wizard-v2/alert-notifications/alert-notifications.component";
 import { AlertTemplate } from "../../components/alert-wizard-v2/alert-template/alert-template.component";
+import { Portal } from "../../components/portal/portal.component";
+import { WizardBottomBar } from "../../components/welcome-onboard-datasource/wizard-bottom-bar/wizard-bottom-bar.component";
 import { PageContentsGridV1 } from "../../platform/components";
-import { AlertEditPageOutletContextProps } from "./alerts-update-page.interfaces";
+import {
+    AlertsSimpleAdvancedJsonContainerPageOutletContextProps,
+    BOTTOM_BAR_ELEMENT_ID,
+} from "../alerts-edit-create-common/alerts-edit-create-common-page.interfaces";
 
 export const AlertsUpdateSimplePage: FunctionComponent = () => {
     const { t } = useTranslation();
+    const [submitBtnLabel, setSubmitBtnLabel] = useState<string>(
+        t("label.update-entity", {
+            entity: t("label.alert"),
+        })
+    );
+    const [isSubmitBtnEnabled, setIsSubmitBtnEnabled] = useState(false);
     const {
         alert,
         handleAlertPropertyChange: onAlertPropertyChange,
@@ -32,14 +43,12 @@ export const AlertsUpdateSimplePage: FunctionComponent = () => {
         selectedAlertTemplate,
         setSelectedAlertTemplate,
         alertTemplateOptions,
-        setShowBottomBar,
-        setIsSubmitBtnEnabled,
-        setSubmitBtnLabel,
-        resetSubmitButtonLabel,
-    } = useOutletContext<AlertEditPageOutletContextProps>();
+        isEditRequestInFlight,
+        handleSubmitAlertClick,
+        onPageExit,
+    } = useOutletContext<AlertsSimpleAdvancedJsonContainerPageOutletContextProps>();
 
     useEffect(() => {
-        setShowBottomBar(true);
         setIsSubmitBtnEnabled(false);
         setSubmitBtnLabel(
             t("message.preview-alert-in-chart-before-submitting")
@@ -47,35 +56,54 @@ export const AlertsUpdateSimplePage: FunctionComponent = () => {
     }, []);
 
     return (
-        <PageContentsGridV1>
-            <Grid item xs={12}>
-                <AlertDetails
-                    alert={alert}
-                    onAlertPropertyChange={onAlertPropertyChange}
-                />
-            </Grid>
-            <Grid item xs={12}>
-                <AlertTemplate
-                    alert={alert}
-                    alertTemplateOptions={alertTemplateOptions}
-                    selectedAlertTemplate={selectedAlertTemplate}
-                    setSelectedAlertTemplate={setSelectedAlertTemplate}
-                    onAlertPropertyChange={onAlertPropertyChange}
-                    onChartDataLoadSuccess={() => {
-                        setIsSubmitBtnEnabled(true);
-                        resetSubmitButtonLabel();
-                    }}
-                />
-            </Grid>
-            <Grid item xs={12}>
-                <AlertNotifications
-                    alert={alert}
-                    initiallySelectedSubscriptionGroups={
-                        selectedSubscriptionGroups
+        <>
+            <PageContentsGridV1>
+                <Grid item xs={12}>
+                    <AlertDetails
+                        alert={alert}
+                        onAlertPropertyChange={onAlertPropertyChange}
+                    />
+                </Grid>
+                <Grid item xs={12}>
+                    <AlertTemplate
+                        alert={alert}
+                        alertTemplateOptions={alertTemplateOptions}
+                        selectedAlertTemplate={selectedAlertTemplate}
+                        setSelectedAlertTemplate={setSelectedAlertTemplate}
+                        onAlertPropertyChange={onAlertPropertyChange}
+                        onChartDataLoadSuccess={() => {
+                            setIsSubmitBtnEnabled(true);
+                            setSubmitBtnLabel(
+                                t("label.update-entity", {
+                                    entity: t("label.alert"),
+                                })
+                            );
+                        }}
+                    />
+                </Grid>
+                <Grid item xs={12}>
+                    <AlertNotifications
+                        alert={alert}
+                        initiallySelectedSubscriptionGroups={
+                            selectedSubscriptionGroups
+                        }
+                        onSubscriptionGroupsChange={onSubscriptionGroupsChange}
+                    />
+                </Grid>
+            </PageContentsGridV1>
+
+            <Portal containerId={BOTTOM_BAR_ELEMENT_ID}>
+                <WizardBottomBar
+                    doNotWrapInContainer
+                    backButtonLabel={t("label.cancel")}
+                    handleBackClick={onPageExit}
+                    handleNextClick={() => handleSubmitAlertClick(alert)}
+                    nextButtonIsDisabled={
+                        !isSubmitBtnEnabled || isEditRequestInFlight
                     }
-                    onSubscriptionGroupsChange={onSubscriptionGroupsChange}
+                    nextButtonLabel={submitBtnLabel}
                 />
-            </Grid>
-        </PageContentsGridV1>
+            </Portal>
+        </>
     );
 };
