+++ thirdeye-ui/src/app/routers/alerts-guided-create/alerts-guided-create.router.interface.ts
@@ -15,7 +15,6 @@
 
 export interface AlertsGuidedCreateRouterProps {
     navigateToAlertDetailAfterSampleAlertCreate?: boolean;
-    useParentForNonWelcomeFlow?: boolean;
     sampleAlertsBottom?: boolean;
     hideSampleAlerts?: boolean;
     hideCurrentlySelected?: boolean;
